+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+  - 💄 [@mixte/snippets] 类型优化: 组件 `AcroDynamicForm` 新增 `input-search:搜索框`、`input-group:输入框组合`、`cascader-panel:级联菜单` 组件支持
 
 ## [v1.13.0], [v1.13.0-beta.2], [v1.13.0-beta.1]
   - 📅 2024-05-30
@@ -7,7 +8,6 @@
   - 💄 [@mixte/snippets] 当组件 `AcroDynamicForm` 从外部传入的 `model` 中某个字段有值时, 那么该字段的 `defaultValue` 不会覆盖原有值
   - 💄 [@mixte/snippets] 类型优化: 组件 `AcroDynamicForm` 字段的 `type` 支持配置为非预设中的字段
   - 💄 [@mixte/snippets] 类型优化: 组件 `AcroDynamicForm` 新增 `year-picker:年份选择器`、`month-picker:月份选择器`、`quarter-picker:季度选择器`、`week-picker:周选择器`、`range-picker:范围选择器`、`input-password:密码输入框` 组件支持
-  - 💄 [@mixte/snippets] 类型优化: 组件 `AcroDynamicForm` 新增  组件支持
   - 🐞 [@mixte/snippets] 修复组件 `AcroDynamicForm` 字段未配置 `defaultValue` 也会写入值的问题
 
 ## [v1.12.0], [v1.12.0-beta.1]